+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v4.0.1
+
+### ✨ Features
+
+* Bean Searcher
+  * 增强 `JsonFieldConvertor` 字段转换器，新增 `catchError` 字段，可配置遇到某些值 JSON 解析异常时，是否自动捕获（即忽略）
+* Bean Searcher Boot Starter
+  * 新增 `bean-searcher.field-convertor.json-catch-error` 配置项，表示是否自动捕获 JSON 解析错误，默认 `false`
+
 # v4.0.0 @ 2023-01-31
 
 ### ✨ Features